Refactor com_atproto::moderation
